Simplify the generator a bit.
